refactor: Rename MoveNet model and add ONNX export wrapper
This commit renames the 'movenet_fpn' module to 'movenet_fpn_sp' to explicitly indicate that this version of the model is designed for single-person pose estimation only. This improves clarity and avoids confusion for future users and developers.

Additionally, a new wrapper has been created for ONNX model export. This new wrapper simplifies the process of exporting the model to the ONNX format, making it easier to deploy the model in various environments and frameworks that support ONNX.